% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_terradat.R
\name{load_terradat}
\alias{load_terradat}
\title{loads TerrADat and LMF tables as sf objects and TerrADatSpeciesIndicators table by default. It can also load TBLSOILPITHORIZONS, TBLSPECRICHDETAIL, TBLSTATESPECIES and POINTWEIGHT tables.}
\usage{
load_terradat(connection, extra = FALSE)
}
\arguments{
\item{connection}{a string of the name of the ODBC connection.  See https://github.com/mschmidty/AimSqlTools for instructions on how to set up connection.}

\item{extra}{a boolean. Default is \code{FALSE}. If \code{TRUE} TBLSOILPITHORIZONS, TBLSPECRICHDETAIL, TBLSTATESPECIES and POINTWEIGHT tables are also loaded.}
}
\description{
loads TerrADat and LMF tables as sf objects and TerrADatSpeciesIndicators table by default. It can also load TBLSOILPITHORIZONS, TBLSPECRICHDETAIL, TBLSTATESPECIES and POINTWEIGHT tables.
}
