% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_terradat.R
\name{query_db2}
\alias{query_db2}
\title{return table from Terrestrial AIM SDE}
\usage{
query_db2(conn, tbl_name, remove_shape = TRUE)
}
\arguments{
\item{conn}{a string of the name of the ODBC connection.  See https://github.com/mschmidty/AimSqlTools for instructions on how to set up connection.}

\item{tbl_name}{a string of the name of the table you wish to query.}

\item{remove_shape}{boolean.  Default \code{TRUE}.  Needed for SDE tables that have ESRI shape information in binary format.  Should be \code{FALSE} if the queried table has no spatial data.}
}
\description{
return table from Terrestrial AIM SDE
}
